Add text for genomic data visualization

Object ccGenomicCellGeom will call the function circlize::circos.genomicText while drawing.

Usage

ccGenomicText(
  region,
  value = NULL,
  y = NULL,
  labels = NULL,
  labels.column = NULL,
  numeric.column = NULL,
  posTransform = NULL,
  direction = NULL,
  facing = "inside",
  niceFacing = FALSE,
  adj = par("adj"),
  cex = 1,
  col = "black",
  font = par("font"),
  padding = 0,
  extend = 0,
  align_to = "region",
  ...
)

Arguments

region

A data frame contains 2 column which correspond to start positions and end positions.

value

A data frame contains values and other information.

y

A vector or a single value indicating position of text.

labels

Labels of text corresponding to each genomic positions.

labels.column

If labels are in value, index of column in value.

numeric.column

Which column in value data frame should be taken as y-value. If it is not defined, only the first numeric columns in value will be taken.

posTransform

Self-defined function to transform genomic positions, see posTransform.default for explanation.

direction

Deprecated, use facing instead.

facing

Passing to circos.text. Settings are similar as col.

niceFacing

Should the facing of text be adjusted to fit human eyes?

adj

Pass to circos.text. Settings are similar as col.

cex

Pass to circos.text. Settings are similar as col.

col

Pass to circos.text. The length of col can be either one or number of rows of region.

font

Pass to circos.text. Settings are similar as col.

padding

pass to posTransform if it is set as posTransform.text.

extend

pass to posTransform if it is set as posTransform.text.

align_to

pass to posTransform if it is set as posTransform.text.

...

Mysterious parameters.

Value

Object ccGenomicCellGeom

Examples

library(circlizePlus)
cc <- ccPlot(initMode = "initializeWithIdeogram", plotType = NULL)
bed <- generateRandomBed(nr = 20)
track1 <- ccGenomicTrack(data = bed, ylim = c(0, 1))
all_chr <- c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9", "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18", "chr19", "chr20", "chr21", "chr22", "chrX", "chrY")
text1 <- ccGenomicText(y = 0.5, labels = "text")
#> Error in ccGenomicText(y = 0.5, labels = "text"): argument "region" is missing, with no default
cells1 <- ccCells(sector.indexes = all_chr) + text1
#> Error: object 'text1' not found
cc + (track1 + cells1)
#> Error: object 'cells1' not found