S4 class ccHeatmap

ccHeatmap is a special class. It can be used not only as a single track but also as the result of adding a heatmap track to a ccPlot

Slots

func

character. Normally it is "circos.heatmap".

params

list. A named list that stores the parameters of the function circlize::circos.heatmap called by the backend.

trackGeoms

list. A list where ccTrackGeom are stored.

cells

list. A list where ccCell are stored.

tracks

list. A list where ccTrack or ccGenomicTrack or ccHeatmap are stored.

links

list. A list where ccLink or ccGenomicLink or ccHeatmapLink are stored.

pars

list. A list where ccPar are stored.