Add points on all sections of a single track.

Object ccTrackGeom will call the function circlize::circos.trackPoints while drawing.

Usage

ccTrackPoints(
  sectors,
  x,
  y,
  pch = par("pch"),
  col = par("col"),
  cex = par("cex"),
  bg = par("bg"),
  factors = sectors
)

Arguments

sectors

A factor or a character vector which represents the categories of data

x

Data points on x-axis

y

Data points on y-axis

pch

Point type

col

Point color

cex

Point size

bg

backgrond color

factors

The same as sectors. It will be removed in future versions.

Value

Object ccTrackGeom

Examples

NULL
#> NULL