circlizePlus: Using ggplot2 feature to write readable R code for circular visualization

circlizePlus is a wrapper for circlize. All components are based on classes and objects. Users can use the addition symbol (+) to combine components for a circular visualization with ggplot2 style.

See also

Author

Maintainer: Tianze Cao hnrcao@qq.com (ORCID)

Authors:

Zheyu Zhang 2054906893@qq.com